requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pyswisseph==2.10.3.2
requests==2.31.0
httpx==0.27.0
timezonefinder==6.2.0
pytz==2024.1
pydantic==2.5.3
google-genai
python-dotenv
Pillow==10.2.0
